+++ development/sabido/lembrete/views.py
@@ -30,8 +30,9 @@ def lembrete_form(request, id = 0):
             form.save()
         return redirect('/lembrete/list')            
 
-def lembrete_delete(id):
-    lembrete = Lembrete.objects.get(pk = id)
-    lembrete.delete()
-    return redirect('/lembrete/list')
+def lembrete_delete(request, id):
+    if request.method == "POST":
+        lembrete = Lembrete.objects.get(pk = id)
+        lembrete.delete()
+        return redirect('/lembrete/list')
